Revert "menu2system"
