+++ TP Final.depend
@@ -3,15 +3,16 @@
 	<stdio.h>
 	<stdlib.h>
 
-1717473293 source:c:\users\bepis\desktop\tp-final\main.c
+1717468395 source:c:\users\bepis\desktop\tp-final\main.c
 	<stdio.h>
 	<stdlib.h>
+	<windows.h>
 	<conio.h>
 
 1717370814 source:c:\users\bepis\desktop\tp-final\song.c
 	"Song.h"
 
-1717473293 c:\users\bepis\desktop\tp-final\song.h
+1717370814 c:\users\bepis\desktop\tp-final\song.h
 	<stdio.h>
 	<stdlib.h>
 
@@ -23,12 +24,12 @@
 	<stdlib.h>
 	"Song.h"
 
-1717473005 source:c:\users\bepis\desktop\tp-final\system.c
+1717383451 source:c:\users\bepis\desktop\tp-final\system.c
 	"System.h"
 	"Song.h"
 	"User.h"
 
-1717473216 c:\users\bepis\desktop\tp-final\system.h
+1717383451 c:\users\bepis\desktop\tp-final\system.h
 	<stdio.h>
 	<stdlib.h>
